feat(coupon): 重构优惠券模块，引入策略模式和工厂模式
- 新增CouponModelFactory统一构建优惠券模型
- 新增DiscountCouponRule和PriceOffCouponRule策略类
- 重命名UserCouponService为CouponService
- 删除UserCouponEvaluator和UserCouponCalculator
- 优化优惠券计算逻辑，支持折扣券和满减券

+++ src/main/kotlin/com/hello/hello_store_service/web/model/view/coupon/CouponResultList.kt
@@ -21,6 +21,6 @@ data class CouponView(
     val promotionCode: String?,
     val promotionSubCode: String?,
     val scopeText: String?,
-    val value: Int,
+    val value: Float,
     val type: Int,
 )
